feat: refactor AppData component and remove AppDataImportModal for improved import/export functionality
